services/fastapi-app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
prometheus-fastapi-instrumentator==6.1.0
httpx==0.25.2
